requirements.txt
streamlit>=1.33
duckdb>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.18
